Change to_i to to_f in calculate_square in application_controller.rb

+++ app/controllers/application_controller.rb
@@ -4,7 +4,7 @@ class ApplicationController < ActionController::Base
   end
 
   def calculate_square
-    @num = params.fetch("user_number").to_i
+    @num = params.fetch("user_number").to_f
     @square_of_num = @num ** 2
     render({ :template => "calculation_templates/square_results.html.erb"})
   end
